En tant qu'utilisateur je souhaite pouvoir me connecter et me déconnecter #3 - feat: refactor footer component and update import path in HomePage
