+++ .prettierrc.json
@@ -4,7 +4,7 @@
   "tabWidth": 2,
   "semi": true,
   "singleQuote": true,
-  "endOfLine": "lf",
+  "endOfLine": "crlf",
   "printWidth": 80,
   "bracketSpacing": true,
   "arrowParens": "always",
